Extract some wrapper tests to separate module
